--- guidance/library/_one_or_more.py
@@ -1,7 +1,0 @@
-from .._guidance import guidance
-from .._grammar import select
-
-
-@guidance(stateless=True)
-def one_or_more(model, value):
-    return model + select([value], recurse=True)
