+++ guidance/library/_tool.py
@@ -1,8 +1,7 @@
 from .._guidance import guidance
 from ._any_char import any_char
 from .._grammar import select, capture, string, commit_point
-from ._zero_or_more import zero_or_more
-from ._one_or_more import one_or_more
+from ._sequences import zero_or_more, one_or_more
 from ._any_char_but import any_char_but
 from ._any_char import any_char
 
@@ -41,8 +40,7 @@ def kwarg():
 def basic_func_grammar(name):
     obj = string(name + "(")
     obj += capture(
-        select([zero_or_more(positional_arg()), ""])
-        + select([zero_or_more(kwarg()), ""]),
+        select([zero_or_more(positional_arg()), ""]) + select([zero_or_more(kwarg()), ""]),
         name="tool_args",
     )
     obj += string(")")
